requirements.txt
# Core ML Libraries
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
scipy==1.11.1

# Advanced ML Models
xgboost==1.7.6
lightgbm==4.0.0
catboost==1.2

# Imbalanced Learning
imbalanced-learn==0.11.0

# Hyperparameter Optimization
optuna==3.3.0
hyperopt==0.2.7

# Model Explainability
shap==0.42.1
lime==0.2.0.1
eli5==0.13.0

# Visualization
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.15.0
plotly-dash==2.13.0

# Web Framework & API
streamlit==1.25.0
fastapi==0.101.1
uvicorn==0.23.2
pydantic==2.1.1
python-multipart==0.0.6

# MLOps & Experiment Tracking
dagshub==0.3.1
mlflow==2.5.0
evidently==0.4.2
great-expectations==0.17.12
dvc==3.22.0

# Data Processing & Storage
joblib==1.3.2
pickle5==0.0.12
h5py==3.9.0
pyarrow==12.0.1

# Web Scraping
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.11.2
scrapy==2.10.0

# Database & Storage
sqlite3
sqlalchemy==2.0.19
pymongo==4.4.1

# Configuration & Utilities
pyyaml==6.0.1
PyYAML==6.0.1
python-dotenv==1.0.0
click==8.1.6
tqdm==4.65.0

# Testing
pytest==7.4.0
pytest-cov==4.1.0
pytest-mock==3.11.1

# Code Quality
black==23.7.0
flake8==6.0.0
isort==5.12.0
mypy==1.5.1

# Jupyter & Notebooks
jupyter==1.0.0
ipykernel==6.25.0
ipywidgets==8.0.7
notebook==7.0.2

# Monitoring & Logging
prometheus-client==0.17.1
psutil==5.9.5
py-cpuinfo==9.0.0

# Time Series (if needed)
statsmodels==0.14.0
prophet==1.1.4

# Additional Utilities
python-dateutil==2.8.2
pytz==2023.3
regex==2023.6.3
wordcloud==1.9.2

# Development Tools
pre-commit==3.3.3
bandit==1.7.5